Move content type to new namespace, add sourcesummary to report sources
