+++ rb_sources/lang/en/rb_source_ojt_evaluation_embedded.php
@@ -1,0 +1,5 @@
+<?php
+
+$string['sourcetitle'] = 'OJT Evaluation Embedded';
+$string['sourcelabel'] = 'OJT Evaluation Embedded';
+$string['sourcesummary'] = 'OJT Evaluation Embedded';
