+++ rb_sources/rb_source_ojt_completion.php
@@ -83,6 +83,7 @@ class rb_source_ojt_completion extends rb_base_source {
         $this->requiredcolumns = $this->define_requiredcolumns();
         $this->sourcetitle = get_string('ojtcompletion', 'rb_source_ojt_completion');
         $this->sourcelabel = get_string('ojtcompletion', 'rb_source_ojt_completion');
+        $this->sourcesummary = get_string('ojtcompletion', 'rb_source_ojt_completion');
         $this->usedcomponents[] = 'mod_ojt';
         $this->usedcomponents[] = 'totara_cohort';
 
@@ -487,117 +488,3 @@ class rb_source_ojt_completion extends rb_base_source {
     }
 
 } // end of rb_source_course_completion class
-
-
-
-/**
- * Restrict content by ojt completion type
- *
- * Pass in an integer that represents a ojt completion type, e.g OJT_CTYPE_TOPIC
- */
-class rb_ojt_completion_type_content extends totara_reportbuilder\rb\content\base {
-
-    /**
-     * Generate the SQL to apply this content restriction
-     *
-     * @param string $field SQL field to apply the restriction against
-     * @param integer $reportid ID of the report
-     *
-     * @return array containing SQL snippet to be used in a WHERE clause, as well as array of SQL params
-     */
-    public function sql_restriction($field, $reportid) {
-        global $CFG, $DB;
-
-        require_once($CFG->dirroot . '/mod/ojt/lib.php');
-
-        // remove rb_ from start of classname
-        $type = substr(get_class($this), 3);
-        $settings = reportbuilder::get_all_settings($reportid, $type);
-
-        return array('base.type = :crbct', array('crbct' => $settings['completiontype']));
-    }
-
-    /**
-     * Generate a human-readable text string describing the restriction
-     *
-     * @param string $title Name of the field being restricted
-     * @param integer $reportid ID of the report
-     *
-     * @return string Human readable description of the restriction
-     */
-    public function text_restriction($title, $reportid) {
-        global $DB;
-
-        // remove rb_ from start of classname
-        $type = substr(get_class($this), 3);
-        $settings = reportbuilder::get_all_settings($reportid, $type);
-
-        return !empty($settings['completiontype']) ? $title.' - '.get_string('type'.$settings['completiontype'], 'ojt') : '';
-    }
-
-
-    /**
-     * Adds form elements required for this content restriction's settings page
-     *
-     * @param object &$mform Moodle form object to modify (passed by reference)
-     * @param integer $reportid ID of the report being adjusted
-     * @param string $title Name of the field the restriction is acting on
-     */
-    public function form_template(&$mform, $reportid, $title) {
-        // get current settings
-        // remove rb_ from start of classname
-        $type = substr(get_class($this), 3);
-        $enable = reportbuilder::get_setting($reportid, $type, 'enable');
-        $completiontype = reportbuilder::get_setting($reportid, $type, 'completiontype');
-
-        $mform->addElement('header', 'ojt_completion_type_header',
-            get_string('showbyx', 'totara_reportbuilder', lcfirst($title)));
-        $mform->setExpanded('ojt_completion_type_header');
-        $mform->addElement('checkbox', 'ojt_completion_type_enable', '',
-            get_string('completiontypeenable', 'rb_source_ojt_completion'));
-        $mform->setDefault('ojt_completion_type_enable', $enable);
-        $mform->disabledIf('ojt_completion_type_enable', 'contentenabled', 'eq', 0);
-        $radiogroup = array();
-        $radiogroup[] =& $mform->createElement('radio', 'ojt_completion_type_completiontype',
-            '', get_string('type'.OJT_CTYPE_OJT, 'ojt'), OJT_CTYPE_OJT);
-        $radiogroup[] =& $mform->createElement('radio', 'ojt_completion_type_completiontype',
-            '', get_string('type'.OJT_CTYPE_TOPIC, 'ojt'), OJT_CTYPE_TOPIC);
-        $mform->addGroup($radiogroup, 'ojt_completion_type_completiontype_group',
-            get_string('includecompltyperecords', 'rb_source_ojt_completion'), html_writer::empty_tag('br'), false);
-        $mform->setDefault('ojt_completion_type_completiontype', $completiontype);
-        $mform->disabledIf('ojt_completion_type_completiontype_group', 'contentenabled',
-            'eq', 0);
-        $mform->disabledIf('ojt_completion_type_completiontype_group', 'ojt_completion_type_enable',
-            'notchecked');
-    }
-
-
-    /**
-     * Processes the form elements created by {@link form_template()}
-     *
-     * @param integer $reportid ID of the report to process
-     * @param object $fromform Moodle form data received via form submission
-     *
-     * @return boolean True if form was successfully processed
-     */
-    public function form_process($reportid, $fromform) {
-        $status = true;
-        // remove rb_ from start of classname
-        $type = substr(get_class($this), 3);
-
-        // enable checkbox option
-        $enable = (isset($fromform->ojt_completion_type_enable) &&
-            $fromform->ojt_completion_type_enable) ? 1 : 0;
-        $status = $status && reportbuilder::update_setting($reportid, $type,
-            'enable', $enable);
-
-        // recursive radio option
-        $recursive = isset($fromform->ojt_completion_type_completiontype) ?
-            $fromform->ojt_completion_type_completiontype : 0;
-        $status = $status && reportbuilder::update_setting($reportid, $type,
-            'completiontype', $recursive);
-
-        return $status;
-    }
-}
-
